requirements.txt
asgiref==3.8.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.26.0
cryptography==35.0.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.11
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-on-heroku==1.1.2
django-rest-knox==4.1.0
djangorestframework==3.15.1
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
stripe==2.61.0
urllib3==1.26.7
